expand env variable inside CYCLONEDDS_URI domain id attribute
