+++ src/main.py
@@ -41,7 +41,7 @@ from PySide6.QtGui import QIcon, QPixmap
 from PySide6.QtQuickControls2 import QQuickStyle
 import logging
 import dds_data
-from dds_service import getConfiguredDomainIds
+from dds_utils import getConfiguredDomainIds
 from models.overview_model import TreeModel, TreeNode
 from models.endpoint_model import EndpointModel
 from models.datamodel_model import DatamodelModel
